#156 refactor chatbox socket code to dedicated file, utilize generic webSocketConnect to standardize connection behavior
